hypnos/requirements.txt
pytest==2.6.3
requests==2.4.3
Flask==0.10.1
Flask-RESTful==0.3.3
Flask-HTTPAuth==2.5.0
itsdangerous==0.24
Jinja2==2.7.3
MarkupSafe==0.23
tornado==4.2
simplejson==3.6.5
Werkzeug==0.10.4
python-dateutil
pika
